# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/jobs/{id}": {
            "get": {
                "tags": [
                    "JobPostings"
                ],
                "summary": "Get a specific job posting",
                "description": "Get a specific job posting.",
                "operationId": "520de9935a90968009e407b27eab9430",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the job posting",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job posting details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "location": {
                                            "type": "string"
                                        },
                                        "employment_type": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "salary_min": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "salary_max": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "currency": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "posted_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "created_by": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job posting not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "JobPostings"
                ],
                "summary": "Update a job posting",
                "description": "Update a specific job posting.",
                "operationId": "6c6685976658720507a8cef80fe6ff54",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the job posting",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "employment_type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Senior Developer"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Kuala Lumpur"
                                    },
                                    "employment_type": {
                                        "description": "Type of employment",
                                        "type": "string",
                                        "enum": [
                                            "Full-time",
                                            "Part-time",
                                            "Contract",
                                            "Internship"
                                        ],
                                        "example": "Full-time"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Job description here."
                                    },
                                    "salary_min": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    },
                                    "salary_max": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 8000
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "MYR"
                                    },
                                    "status": {
                                        "description": "Job status",
                                        "type": "string",
                                        "enum": [
                                            "open",
                                            "closed"
                                        ],
                                        "example": "open"
                                    },
                                    "posted_at": {
                                        "description": "Posting date (Y-m-d H:i:s), must be in the future",
                                        "type": "string",
                                        "example": "2025-11-30 16:46:32"
                                    },
                                    "expires_at": {
                                        "description": "Expiration date (Y-m-d H:i:s)",
                                        "type": "string",
                                        "example": "2025-12-10 16:46:32"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job posting updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job posting updated successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Senior Developer"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Kuala Lumpur"
                                                },
                                                "employment_type": {
                                                    "type": "string",
                                                    "example": "Full-time"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Job description here."
                                                },
                                                "salary_min": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 5000
                                                },
                                                "salary_max": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 8000
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "MYR"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "open"
                                                },
                                                "posted_at": {
                                                    "type": "string",
                                                    "example": "2025-11-08 16:46:32"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "example": "2025-11-15 16:46:32"
                                                },
                                                "created_by": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Job posting not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}